requirements.txt
fastapi==0.115.12
uvicorn==0.34.0
python-dotenv==1.0.1
langchain>=1.2.0,<2.0.0
langchain_openai>=1.1.0,<2.0.0
langchain_google_genai>=3.0.0,<4.0.0
langchain-groq>=1.0.0,<2.0.0
langchain-anthropic>=1.3.0,<2.0.0
langchain_community>=0.4.0,<1.0.0
langchain-astradb>=1.0.0,<2.0.0
langgraph>=1.0.0,<2.0.0
requests>=2.32.5,<3.0.0
pymupdf==1.25.1
scikit-learn==1.6.0
openpyxl==3.1.5
streamlit==1.41.1
SQLAlchemy==2.0.36
# mysqlclient==2.2.6  # Commented out - using pymysql instead (pure Python)
pymysql==1.1.1
aiomysql==0.2.0
bs4==0.0.2
playwright==1.49.1
w3lib==2.2.1
extruct==0.18.0
redis==5.2.1
boto3==1.38.38
pypdf==5.6.0
mistralai==1.8.2
PyPDF2==3.0.1
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
